Fix type and test issues
